ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pillow==10.1.0
# numpy==1.25.2
opencv-python==4.8.1.78
# torch==2.1.1
# torchvision==0.16.1
transformers==4.35.2
segment-anything==1.0
ultralytics==8.3.0
# ultralytics==8.0.202
groq==0.4.1
reportlab==4.0.7
aiofiles==23.2.0
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2
requests==2.31.0
tensorflow==2.15.0
boto3==1.34.0